# swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "SeqUI Server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/dashboards/v1/{uuid}": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "dashboards_v1"
                ],
                "operationId": "dashboards_v1_getByUUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/dashboards.v1.Dashboard"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "dashboards_v1"
                ],
                "operationId": "dashboards_v1_delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response"
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "dashboards_v1"
                ],
                "operationId": "dashboards_v1_update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dashboard UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dashboards.v1.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response"
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
        "/massexport/v1/cancel": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "massexport_v1"
                ],
                "operationId": "massexport_v1_cancel",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/massexport.v1.CancelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response"
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
        "/massexport/v1/start": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "massexport_v1"
                ],
                "operationId": "massexport_v1_start",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/massexport.v1.StartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/massexport.v1.StartResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
        "/seqapi/v1/aggregation": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_getAggregation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetAggregationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetAggregationResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "UnexpectedError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dashboards.v1.Dashboard": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string"
                }
            }
        },
...
        "dashboards.v1.UpdateRequest": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "massexport.v1.CancelRequest": {
            "type": "object",
            "properties": {
                "session_id": {
                    "type": "string"
                }
            }
        },
...
        "massexport.v1.StartRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "name": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "window": {
                    "type": "string"
                }
            }
        },
        "massexport.v1.StartResponse": {
            "type": "object",
            "properties": {
                "session_id": {
                    "type": "string"
                }
            }
        },
        "seqapi.v1.Aggregation": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationBucket"
                    }
                },
                "not_exists": {
                    "type": "integer"
                }
            }
        },
        "seqapi.v1.AggregationBucket": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "not_exists": {
                    "type": "integer"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "seqapi.v1.AggregationFunc": {
            "type": "string",
            "enum": [
                "count",
                "sum",
                "min",
                "max",
                "avg",
                "quantile",
                "unique"
            ],
            "x-enum-varnames": [
                "afCount",
                "afSum",
                "afMin",
                "afMax",
                "afAvg",
                "afQuantile",
                "afUnique"
            ]
        },
        "seqapi.v1.AggregationQuery": {
            "type": "object",
            "properties": {
                "agg_func": {
                    "default": "count",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.AggregationFunc"
                        }
                    ]
                },
                "field": {
                    "type": "string"
                },
                "group_by": {
                    "type": "string"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "seqapi.v1.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "default": "ERROR_CODE_NO",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.ErrorCode"
                        }
                    ]
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "seqapi.v1.ErrorCode": {
            "type": "string",
            "enum": [
                "ERROR_CODE_NO",
                "ERROR_CODE_PARTIAL_RESPONSE",
                "ERROR_CODE_QUERY_TOO_HEAVY",
                "ERROR_CODE_TOO_MANY_FRACTIONS_HIT"
            ],
            "x-enum-varnames": [
                "aecNo",
                "aecPartialResponse",
                "aecQueryTooHeavy",
                "aecTooManyFractionsHit"
            ]
        },
...
        "seqapi.v1.GetAggregationRequest": {
            "type": "object",
            "properties": {
                "aggField": {
                    "type": "string"
                },
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationQuery"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "query": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "seqapi.v1.GetAggregationResponse": {
            "type": "object",
            "properties": {
                "aggregation": {
                    "$ref": "#/definitions/seqapi.v1.Aggregation"
                },
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.Aggregation"
                    }
                },
                "error": {
                    "$ref": "#/definitions/seqapi.v1.Error"
                },
                "partialResponse": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "description": "Authentication token, prefixed by Bearer: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}